# services/tenant/openapi.yaml
openapi: 3.1.0
info:
  title: Tenant
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: API Gateway
tags:
- name: tenant-controller
  description: "Tenant lifecycle: register (install) and uninstall Forge app tenants"
paths:
  /api/{version}/tenants:
    post:
      tags:
      - tenant-controller
      summary: Register a tenant
      description: "Registers a new tenant installation or acknowledges a reinstall/redelivery.\
        \ Returns 201 on first install with a Location header, 200 on reinstall."
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTenantRequest"
        required: true
      responses:
        "201":
          description: Tenant registered (first install)
          headers:
            Location:
              description: URI of the newly created tenant resource
              style: simple
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantResponse"
              examples:
                created:
                  description: created
                  value:
                    installationId: install-xyz-456
                    appId: app-1
                    appVersion: 1.0.0
                    environmentId: env-1
                    siteUrl: https://example.atlassian.net
                    installerAccountId: installer-1
                    status: ACTIVE
                    installedAt: 2025-01-15T10:30:00Z
                    updatedAt: 2025-01-15T10:30:00Z
                    uninstalledAt: null
                    purgeAfter: null
        "200":
          description: Tenant already exists (reinstall/redelivery)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantResponse"
              examples:
                reinstall:
                  description: reinstall
                  value:
                    installationId: install-new-789
                    appId: app-1
                    appVersion: 2.0.0
                    environmentId: null
                    siteUrl: null
                    installerAccountId: null
                    status: ACTIVE
                    installedAt: 2025-01-15T10:30:00Z
                    updatedAt: 2025-06-15T10:30:00Z
                    uninstalledAt: null
                    purgeAfter: null
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              examples:
                validationError:
                  summary: Validation failure
                  description: validationError
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: The request body failed validation
                    code: VALIDATION_ERROR
                    traceId: 6d3e5f1a2b4c7d8e9f0a1b2c3d4e5f6a
                    errors:
                    - field: id
                      code: REQUIRED
                      message: must not be blank
                malformedRequest:
                  summary: Malformed JSON body
                  description: malformedRequest
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: The request body could not be read
                    code: MALFORMED_REQUEST
                    traceId: 6d3e5f1a2b4c7d8e9f0a1b2c3d4e5f6a
                missingTenantContext:
                  summary: Missing X-Tenant-ID header
                  description: missingTenantContext
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: Tenant context is required but was not provided
                    code: MISSING_TENANT_CONTEXT
                    traceId: 6d3e5f1a2b4c7d8e9f0a1b2c3d4e5f6a
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    delete:
      tags:
      - tenant-controller
      summary: Uninstall a tenant
      description: Records a Forge app uninstall. Marks the tenant as UNINSTALLED
        and schedules purge. Idempotent for already-uninstalled tenants.
      operationId: uninstall
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UninstallTenantRequest"
      responses:
        "200":
          description: Tenant uninstalled (or already uninstalled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UninstallTenantResponse"
              examples:
                uninstalled:
                  description: uninstalled
                  value:
                    installationId: install-xyz-456
                    appId: app-1
                    appVersion: 1.0.0
                    environmentId: null
                    siteUrl: null
                    installerAccountId: null
                    status: UNINSTALLED
                    installedAt: 2025-01-15T10:30:00Z
                    updatedAt: 2025-06-15T10:30:00Z
                    uninstalledAt: 2025-06-15T10:30:00Z
                    purgeAfter: 2025-07-15T10:30:00Z
        "400":
          description: Missing tenant context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              examples:
                missingTenantContext:
                  description: missingTenantContext
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: Tenant context is required but was not provided
                    code: MISSING_TENANT_CONTEXT
                    traceId: 6d3e5f1a2b4c7d8e9f0a1b2c3d4e5f6a
        "404":
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              examples:
                notFound:
                  description: notFound
                  value:
                    type: about:blank
                    title: Tenant Not Found
                    status: 404
                    detail: No tenant exists for the given cloud ID
                    code: TENANT_NOT_FOUND
                    traceId: 6d3e5f1a2b4c7d8e9f0a1b2c3d4e5f6a
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    parameters:
    - name: version
      in: path
      required: true
      schema:
        type: integer
        format: int32
      example: 1
components:
  schemas:
    App:
      type: object
      description: Application metadata for the installation
      properties:
        id:
          type: string
          description: Unique application identifier
          example: app-1
          minLength: 1
        version:
          type: string
          description: Application version
          example: 1.0.0
        name:
          type:
          - string
          - "null"
          description: Human-readable application name
          example: My App
        ownerAccountId:
          type:
          - string
          - "null"
          description: Account ID of the application owner
          example: owner-abc-123
      required:
      - id
      - version
    Environment:
      type: object
      description: Environment context for the installation
      properties:
        id:
          type:
          - string
          - "null"
          description: Environment identifier
          example: env-prod-1
    RegisterTenantRequest:
      type: object
      description: Request body for registering a tenant installation
      properties:
        id:
          type: string
          description: Unique installation identifier
          example: install-xyz-456
          minLength: 1
        installerAccountId:
          type:
          - string
          - "null"
          description: Account ID of the user who installed the app
          example: account-abc-123
        app:
          $ref: "#/components/schemas/App"
        environment:
          type: "null"
          $ref: "#/components/schemas/Environment"
        siteUrl:
          type:
          - string
          - "null"
          description: Base URL of the Atlassian site
          example: https://mysite.atlassian.net
      required:
      - app
      - id
    TenantResponse:
      type: object
      description: Tenant registration result
      properties:
        installationId:
          type: string
          description: Installation identifier
          example: install-xyz-456
        appId:
          type: string
          description: Application identifier
          example: app-1
        appVersion:
          type:
          - string
          - "null"
          description: Application version
          example: 1.0.0
        environmentId:
          type:
          - string
          - "null"
          description: Environment identifier
          example: env-1
        siteUrl:
          type:
          - string
          - "null"
          description: Atlassian site URL
          example: https://example.atlassian.net
        installerAccountId:
          type:
          - string
          - "null"
          description: Account ID of the installer
          example: installer-1
        status:
          type: string
          description: Current tenant status
          enum:
          - ACTIVE
          - SUSPENDED
          - UNINSTALLED
          example: ACTIVE
        installedAt:
          type: string
          format: date-time
          description: Timestamp of initial installation
          example: 2025-01-15T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of last update
          example: 2025-01-15T10:30:00Z
        uninstalledAt:
          type:
          - string
          - "null"
          format: date-time
          description: Timestamp when the tenant was uninstalled
          example: 2025-06-15T10:30:00Z
        purgeAfter:
          type:
          - string
          - "null"
          format: date-time
          description: Timestamp after which tenant data may be purged
          example: 2025-07-15T10:30:00Z
    Violation:
      type: object
      description: Field-level validation error
      properties:
        field:
          type: string
          description: Request field that failed validation
          example: app.id
        code:
          type: string
          description: Machine-readable violation category
          enum:
          - REQUIRED
          - INVALID_FORMAT
          - TOO_SHORT
          - TOO_LONG
          - INVALID_VALUE
          example: REQUIRED
        message:
          type:
          - string
          - "null"
          description: Server-localized human-readable message
          example: must not be blank
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Details response body
      properties:
        type:
          type: string
          description: URI reference identifying the problem type
          example: about:blank
        title:
          type: string
          description: Short human-readable summary of the problem
          example: Bad Request
        status:
          type: integer
          format: int32
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: The request body failed validation
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        traceId:
          type: string
          description: Distributed trace identifier for correlation
          example: 6d3e5f1a2b4c7d8e9f0a1b2c3d4e5f6a
        errors:
          type: array
          description: Field-level validation errors (present when code is VALIDATION_ERROR)
          items:
            $ref: "#/components/schemas/Violation"
    UninstallTenantRequest:
      type: object
      description: Forge preUninstall life-cycle event payload
      properties:
        id:
          type: string
          description: Installation identifier
          example: install-xyz-456
        app:
          type: "null"
          $ref: "#/components/schemas/App"
          description: Application metadata
    UninstallTenantResponse:
      type: object
      description: Tenant uninstall result
      properties:
        installationId:
          type: string
          description: Installation identifier
          example: install-xyz-456
        appId:
          type: string
          description: Application identifier
          example: app-1
        appVersion:
          type:
          - string
          - "null"
          description: Application version
          example: 1.0.0
        environmentId:
          type:
          - string
          - "null"
          description: Environment identifier
          example: env-1
        siteUrl:
          type:
          - string
          - "null"
          description: Atlassian site URL
          example: https://example.atlassian.net
        installerAccountId:
          type:
          - string
          - "null"
          description: Account ID of the installer
          example: installer-1
        status:
          type: string
          description: Current tenant status
          enum:
          - ACTIVE
          - SUSPENDED
          - UNINSTALLED
          example: UNINSTALLED
        installedAt:
          type: string
          format: date-time
          description: Timestamp of initial installation
          example: 2025-01-15T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of last update
          example: 2025-06-15T10:30:00Z
        uninstalledAt:
          type:
          - string
          - "null"
          format: date-time
          description: Timestamp when the tenant was uninstalled
          example: 2025-06-15T10:30:00Z
        purgeAfter:
          type:
          - string
          - "null"
          format: date-time
          description: Timestamp after which tenant data may be purged
          example: 2025-07-15T10:30:00Z